requirements.txt
# Core data engineering
pyspark==3.5.0
pandas==2.1.4
pyarrow==14.0.1
numpy==1.26.2

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Data validation
great-expectations==0.18.8

# HTTP + data download
requests==2.31.0
tqdm==4.66.1

# Environment management
python-dotenv==1.0.0

# SQL + database
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.14          # Mock AWS services for testing

# Utilities
loguru==0.7.2          # Better logging
click==8.1.7           # CLI interface
pyyaml==6.0.1

# Notebook (optional, for exploration)
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
